+++ app/api/users/route.ts
@@ -6,6 +6,6 @@ export const dynamic = "force-dynamic";
 
 export async function POST(req: NextRequest) {
   const body = (await req.json()) as User;
-  const db = createUser(body);
+  const db = await createUser(body);
   return NextResponse.json(db);
 }